fix ut location
